CP init scripts to jenkins-init-scripts
This is a transition patch. Global-JJB now provides
jenkins-init-scripts and we need to switch to using them.
OpenDaylight has a few custom components not covered by global-jjb
so we are using the local-init.sh entrypoint to add custom ODL code.

Once this is merged we can test on the sandbox and if successful
remove everything in the jenkins-scripts directory.

Change-Id: I5bbdb4e06aabd062226124a7e02c9d187ac6df0d
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>