Update script to deploy combo node
This updates the openstack-run.sh script to allow setting
up the combo deployment where there is only one node that
is a control+compute all in one.

The design reuses the majority of the existing design where
there are control node configuration and then compute node
configuration. A combo node is really a control node with
n-cpu configuration. Therefore, the updates are just to the
control node configuration to add the n-cpu configuration
when the combo node is selected via the job yaml.

JIRA: NETVIRT-1370
Change-Id: I8cfda38708575646095ae2878d64fb68b4359f03
Signed-off-by: Sam Hague <shague@redhat.com>