Update patch set 3
Patch Set 3:

> Are we sure this change is needed?
 > I think, all the type_drivers are loaded by default

Venkat, don't we have enough logs in our jobs to be able to
verify this? I know Srinivas was working on this because of
a problem in the sandbox where a gre network create failed
because of no tenant network available.

Patch-set: 3
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0