Add default script and config plan option
This is a continuation of I6c594be8ac9b084f18171dc8ab4605b4062dfd3f
where we add the defaults for other jobs types. Currently no project
other than netvirt defines the project.txt file so this is a noop.
ovsdb will be using the option in a dependent patch to re-enable the
jvm plots.

Change-Id: I0e967b239f1645a1cf2013a1d646656ddc1ef999
Signed-off-by: Sam Hague <shague@redhat.com>