Simplify jenkins-jobs execution
The idea here is instead of having folks install JJB manually by
running local virtualenv commands and ensuring their version of
JJB matches what's defined in releng-jobs.yaml. We can use tox
instead.

Example:

    tox -e jenkins-jobs test jjb/
    tox -e jenkins-jobs -- test jjb/

This takes care of pulling in the right JJB version into a tox
managed virtualenv. If the version is upgraded then tox will
automatically update this version as well. Should save some
headaches for Sandbox testers in the future when we upgrade
JJB.

Also add a check to ensure that the versions in the 2 files always
match. Otherwise throw an error.

Change-Id: I509fe1fd4f0edd4e7779b0a1b082626c2d7db610
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>