Update patch set 4
Patch Set 4: Code-Review+2

not sure if it holds up across all suites, but comparing the runtime
of just the robot code in the 1cmb vs 1cnt-2cmp jobs, we have 
~20% time savings (9m vs 11m) in the l2.suite.

if it does hold up, and our other suites actually map properly to the
1cmb job, then the 2 hours runtime would shrink to near 90m. just
data. not anything to worry about at the moment.

Patch-set: 4
Label: Code-Review=+2
Label: SUBM=+1