Update patch set 2
Patch Set 2:

The idea is to update Venkat's patch with these changes. This patch add support for multiple OVS versions and simplifies the build.
I think we should move it to packer as soon as possible through, as the build may interfere with what is already there, mostly on system wide python packages installd via pip.

Patch-set: 2