Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

I'm not a fan of completely duplicating global-jjb just to add a single script run at the end of the build.

Why not create a new job that is configured to run after the verify job or merge job and reuses the artifacts from the previous job in order to run the test scripts.

Alternatively we can work with upstream global-jjb to add the ability for these jobs to allow projects to add a post build script to run and perform additional testing.

Patch-set: 2
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1