Update patch set 5
Patch Set 5:

what do you guys think now? i had lots of problems trying to
do everything in a single venv. so, basically I did the pip3
installs first at the system level, then did the venv for python2
for the rest of the stuff.

an ovsdb csit job passed with it like this.

Patch-set: 5