Remove validate-job for mdsal neon stream
The validate autorelease job was not designed to run on projects
that have different branch name then all the other projects. This
causes it to try to clone the wrong autorelease branch when running.

Considering that mdsal is becoming a MRI project soon in the Neon
stream. Let's remove the validate-autorelease job. The autorelease
nightly build will catch MD-SAL issues anyway if there is a bug.

Change-Id: Ia9f1c73833bb70af8f01b68d59ccb6f9adb9f2ba
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>