Use common-functions for deploy scripts
Next step would be to split the
integration-deploy-controller-run-test.sh into a configure step and
a run step. With that the configure step means that the same script
can be used for both deploy jobs. Notice that now the
integration-configure-clustering.sh script is exactly the same as
the beginning of the integration-deploy-controller-run-test.sh so
the same file can be used in the job templates.

Change-Id: I572fdb1480b6f38f6fee8792bb2bd0b908cdf752
Signed-off-by: Sam Hague <shague@redhat.com>