Make merge jobs run daily
Back on Rackspace we had unstable cloud and we put in many measures
to reduce the amount of jobs run so that we could try to keep the
cloud as stable as possible. With Vexxhost the cloud is much more
stable. Running merge job daily should be fine at this point. Let's
bump it to a daily run so that projects have more chances to have
a passing merge job before the 3 week SNAPSHOT cleanup cron in Nexus.

Change-Id: I7e47bdfa33f053997b59d51abc21809e0f20cb23
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>