Update patch set 2
Patch Set 2: Code-Review-2

we need to understand what this could break. I thought this was
in a spin-up script, but this is in the code that builds our
images. I'm guessing our existing images might still be old enough
that they still have a working python-six installed and whatever
is using it will break if we update images and don't have it.

Venkat, can you help us here?

Patch-set: 2