Enhance the RPM CI scripts
This patch removes the common parts with in the CI scripts.

Also it fixes CI: For logging configuration, only console type
is covered by the tests. However, the default type is file, for
which there are no tests, and the console tests will fail, which
are triggered automatically. For now it hard-codes the logging
mechanism to type console.

Change-Id: I3883d1307cc044d672e18eb7b41ff9047b728108
Signed-off-by: Taseer <taseer94@gmail.com>