Update patch set 3
Patch Set 3:

> No problem in being extra careful but just in case it was missed:
 > this should only really affect when networking-sfc plugin is
 > enabled, which is only for the SFC jobs.

yeah, I noticed and even double checked the same. It's not much
overhead to push a job and run it once in the sandbox (maybe 5m
time) and just wait until it's done.

appreciate it. We aren't perfect, but really try to keep our jobs
as stable as possible and this is one way. :)

Patch-set: 3
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0