Update patch set 14
Patch Set 14:

> This doesn't work. It's breaking CI with:
 > 
 > + sudo ansible-playbook -i localhost, -c local /w/workspace/ansible-test-rpm-master/ansible/examples/rpm_8_devel_odl_api.yml
 > --extra-vars @/w/workspace/ansible-test-rpm-master/ansible/examples/log_vars.json
 > ERROR! Unable to retrieve file contents
 > Could not find or access '/w/workspace/ansible-test-rpm-master/ansible/examples/log_vars.json'
 > on the Ansible Controller.
 > 
 > I think the issue is that you removed the common parts as we
 > needed, but didn't extract them to a common builder script and add
 > them back. So basically just removed the logic without adding it
 > back anywhere. Need to create a new builder script that owns this
 > common setup and run it as the first builder step, before these
 > test-specific scripts. So, here:
 > 

So I guess this actually isn't why it's breaking, although it's still feedback I'd like to see addressed. It was removed in two duplicate places, but still exists in the first builder called. So that works, but I'd still much rather see it extracted to a dedicated "setup" script that runs first, so there's not some hidden ordering of the test builder scripts.

Patch-set: 14