Clarify log message to not look like a failure
The result of checking if there's a known_host file we should clean up
looks a lot like a failure in the logs. It's actually not a big deal.

+ rm /home/jenkins/.ssh/known_hosts
rm: cannot remove /home/jenkins/.ssh/known_hosts: \
  No such file or directory
+ echo 'No known_hosts file'
No known_hosts file

This is immediately followed by an obvious transition in the logs,
enforcing the idea that this 'error' caused the transition.

The linked Jira was raised about this, thinking it was an real error.

Change-Id: I7baa6a112ecdfc44d3511accfe5624d7572ba713
Jira: INTPAK-201
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>