Speed up RPM verify job, rely more on verify full
The current rpm-verify and rpm-verify-full jobs are almost the same. The
idea is to have one quick that runs on every patchset and one
full-coverage that can be triggered by a keyword when wanted. Remove
some coverage from verify to speed it up. Same tests are still in full.

Change-Id: I66b4857a71d713f606d899843fa99a79f2c4a65d
Jira: INTPAK-212
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>