Update patch set 1
Patch Set 1:

> @jamo @skitt for future reference global-jjb jobs also support a
 > "disable-job: true|false" parameter now.
 > 
 > Simply add it to a project section and it will disable all of the
 > jobs instead of deleting.

right, but in the past I remember being discouraged from doing this
disable trick because of some reason. can't quite remember, but maybe
it was that once we disable it, we have to keep the flag, whereas
if we never disable it to start with, we never need the flag. ring
a bell?

Patch-set: 1