Update patch set 1
Patch Set 1:

> > @jamo @skitt for future reference global-jjb jobs also support a
 > > "disable-job: true|false" parameter now.
 > >
 > > Simply add it to a project section and it will disable all of the
 > > jobs instead of deleting.
 > 
 > right, but in the past I remember being discouraged from doing this
 > disable trick because of some reason. can't quite remember, but
 > maybe
 > it was that once we disable it, we have to keep the flag, whereas
 > if we never disable it to start with, we never need the flag. ring
 > a bell?

Yeah it's because when you remove the section Jenkins keeps the previous setting rather than resetting the setting to default. The newer versions of Jenkins no longer does this.

Regardless of that though, this is a setting inside global-jjb templates so the default is always "true" if you are using a global-jjb template so even if we were on an old version of Jenkins the right setting will always get set. "disable-job: true" is different than the one your thinking which was simply just "disabled: true" from JJB.

Patch-set: 1