Use ports websocket URI for haproxy healthcheck
port 8185 is up and listening even before the neutron
data path is registered, so just checking on the port
is not enough.

also, checking on /diagstatus from port 8181 is not
helping much either, since that also will tell us
nothing about whether the websocket is going to work
for networking_odl.

Change-Id: Ie37f7a68f1d4785a7037f1490293bfb618de7bab
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>