Update patch set 10
Patch Set 10: Code-Review+1

(2 comments)

this is working as expected in the sandbox.

when the block_on_csit is set to false, and the csit job fails,
we wont be failing the parent distro-check job. Conversely, when
block_on_csit is set to true with the failure-threshold set to
UNSTABLE, we will fail the distro-check job.

I also tested both cases with the csit job passing, just to make
sure we got blue dots in distro-check.

Patch-set: 10
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1