Move streamtestplan variable out of defaults.yaml
The problem with nesting variables in defaults.yaml is it causes
everyone, regardless of if they care about streamtestplan to
require the variables defined in streamtestplan to be required in
their templates. This is causes job-templates that never needed
these variables to have to define a dummy variable in order to
successfully pass.

Considering no existing csit jobs override the 'streamtestplan'
variable, remove it as a option and code the pattern into the
job-templates that use it.

Change-Id: Ib2355753031a908aafa7e43bcc01c34cce770e93
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>