Move scriptplan variable out of defaults.yaml
The problem with nesting variables in defaults.yaml is it causes
everyone, regardless of if they care about scriptplan to
require the variables defined in scriptplan to be required in
their templates. This is causes job-templates that never needed
these variables to have to define a dummy variable in order to
successfully pass.

Considering no existing csit jobs override the 'scriptplan'
variable, remove it as an option and code the pattern into the
job-templates that use it.

Change-Id: I9244132a02f85d69806988c1faffa2760b924667
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>