Move configplan variable out of defaults.yaml
The problem with nesting variables in defaults.yaml is it causes
everyone, regardless of if they care about configplan to
require the variables defined in configplan to be required in
their templates. This is causes job-templates that never needed
these variables to have to define a dummy variable in order to
successfully pass.

Considering no existing csit jobs override the 'configplan'
variable, remove it as an option and code the pattern into the
job-templates that use it.

Change-Id: I2e0c5d655de12b1f274da282ab66cd62e2a02dd0
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>