Update patch set 1
Patch Set 1:

Only thing that I'm a little unclear about is if the "robot" publisher and the "plot" publisher depend on each other for ordering. On some CSIT jobs we declare a bunch of plot configuration but I'm not sure how the CSVs it uses are getting generated.

Can someone on integration confirm or show me some example jobs where plot is being used. As far as I can tell it's only plotting pass/fail on the job build?

Patch-set: 1
Label: Verified=0