Fix SC2059 for ShellCheck 0.4.4, printf formatting
Consistently with shellcheck 0.4.4, but tested-not 0.4.5, 0.4.6 or
0.5.0, linting fails on inserting strings directly into printf first
arguments vs passing as secondary arguments. The reasoning is related to
vars with special chars being executed vs treated literally.

The default version of shellcheck in still-supported OS distro package
repos (for example Debian 9.6) is 0.4.4. The packer baseline script
installs shellcheck 0.4.6 manually. Shellcheck 0.5.0 is the current
latest.

Change-Id: I202611873b7bd09c65061e29eb2d1bfbdb2085c1
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>