Update patch set 3
Patch Set 3:

> Jamo/Sam? This touches your OpenStack jobs.

yeah, I looked at this originally but then got scared to just
merge it when I saw which files it was touching. Daniel, if you
have time to create a netvirt csit job in the sandbox based on
this patch, and it runs fine (which I expect) then I'll merge.

I'll be able to work on that next week, if needed.

Patch-set: 3
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0