Add blank TOOLS_SYSTEM_IP var to pybot cmdline
When running the full list of netvirt suites with the apex
snapshots, the l2l3_gateway_mac_arp suite sources a bgp
robot lib which in turn tries to reference the TOOLS_SYSTEM_IP.
This is not needed or used in any of the suites, but because
it's sourced but not initialized (even to null) we will get
an ugly error in the robot html report.

looks something like:

ERROR    Error in file <file> Variable '${TOOLS_SYSTEM_IP}' not found.

adding it here as a null value will remove that error

Change-Id: I0bda42e3127cb38f943a7b5b014f13a88da194bd
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>