Cleanup staging repo credentials from .netrc
Patch Iac41cc915d810b5289a96f27d4368c8abf434152 forgot to cleanup
after itself and causes issues when logs script comes up because
of duplicate entry in .netrc. This in turn makes the cURL command
in the logs script grab the first entry which is incorrect for
log shipping.

Change-Id: Ic00fc104d3dbd5648cde261627cad26068b249fb
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>