Ignore intg/dist/opendaylight in version bump
The change Ic6e6bb737af9a8a3035c2ee4d4b590072073bf46 does not ignore
the absense of the intg/dist dir and instead fails the version bump
when the directory is not available. Instead ignore if this dir is
not present since the opendaylight/pom.xml builds the common
distribution (managed distribution + SM projects) for Fluorine onwards.

Change-Id: Ia2c0d0ae3b9eec1491f60e8bcbce71fc61ad26cd
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>