Increase patch-test timeout 360m->720m
Even with 360m for the patch test timeout, we sometimes
will hit that 6 hour limit before the parent gets the
results back from the csit job it triggers. In the case
of netvirt csit jobs, those run for 3-4 hours on their
own, and if the jenkins queue is large that 6h timeout
is easy to hit. Additionally, we have recently reduced
the number of parallel csit jobs from 25 to 20 [0] so
it's even more likely for us to have a larger jenkins
queue.

[0] https://git.opendaylight.org/gerrit/#/c/78776/

Change-Id: Iab733dee2dbac5d30a3bdad228069aa4bd122bf1
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>