Update patch set 6
Patch Set 6:

You can see the dist test change in action in: https://jenkins.opendaylight.org/sandbox/job/integration-distribution-test-fluorine/4
PS- The unstable result is not because this change, it already exists and it is due to setuptools module version conflict in the script that creates csit_failed_tests.txt. Thanh, any idea why is this failing?

Patch-set: 6