Update patch set 1
Patch Set 1:

I *think* the netconf job we were watching was more stable after
reducing our parallel job count, although it still did have a few
random failures. So the question is, was it because of this
reduction of parallel jobs? If so, then would reducing the number
of jobs even more make that job 100% stable?

if so, what would that mean?

either way, I think figuring out the root cause would be a giant
task. So, if we are ok with the unstable netconf job maybe this
patch is ok.

Patch-set: 1