Update patch set 3
Patch Set 3:

So far JDKVERSION parameter is not passed from AR to distribution-test, JDKVERSION is normally set in the CSIT job itself and today's default is openjdk8. We can set the JDKVERSION in CSIT for a given release by just editing the projects csit yaml. Or we could more dynamically pass this parameter from AR by making a simple change in AR and distribution-test jobs.

Patch-set: 3