Update patch set 3
Patch Set 3:

> So far JDKVERSION parameter is not passed from AR to
 > distribution-test, JDKVERSION is normally set in the CSIT job
 > itself and today's default is openjdk8. We can set the JDKVERSION
 > in CSIT for a given release by just editing the projects csit yaml.
 > Or we could more dynamically pass this parameter from AR by making
 > a simple change in AR and distribution-test jobs.

JDK_VERSION should be set only once when the VM is being started and can be passed from the AR job. I'll work on that as a separate change, since there are some changes to the way we trigger the downstream csit jobs or this canbe modified here to also pass the {java-version} as an additional param in L192.

https://git.opendaylight.org/gerrit/#/c/79799/6/jjb/autorelease/autorelease-templates.yaml

Patch-set: 3
Reviewer: Gerrit User 6258 <6258@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0