Disable pkg jobs that require substantial upkeep
Disable the Int/Pack package building tests that require the most
maintenance. Some tests require URLs to actual artifacts that
fairly-frequently expire, requiring the tests to be updated with new
URLs when they do. This will reduce test coverage somewhat, but I think
it's currently worth it for the lower maintenance load.

Change-Id: I04fb222fa00ed9e5ae06e9350f819e28bb3c9ae9
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>