Test job for artifact release automation
This job should trigger when adding a pom to
the zzz-test-release project.

It will create the staging artifact and
logs I need to automate releasing  said artifact,
which I will fist test in the sandbox.

Releasing the artifact will fail but at least we will see
if its trying to do the right thing.

Change-Id: I1ef348a4aba57c43963812fbeb000f53d956790b
Signed-off-by: Aric Gardner <agardner@linuxfoundation.org>