Update Ubuntu 16.04 mininet-ovs-26 images
Some of the working images on prod werer removed because
the openstack cron job did not run for a few weeks on
stack removal failures.

All new images are to be marked as protected, when we update
new images, if the cron job is not running then the imaege
might get removed on the infra.

https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/builder-openstack-cron/19027/console.log.gz

Change-Id: I2a77a84a11c630274dd9cc2cabb76387878d79d6
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>