Update CentOS 7 devstack-rocky image to latest
Some of the working images on prod were removed because
the openstack cron job did not run for a few weeks on
stack removal failures.

All new images are to be marked as protected, when we update
new images, if the cron job is not running then the imaege
might get removed on the infra.

https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/builder-openstack-cron/19027/console.log.gz

Change-Id: I7b52aa722a24ba0efd61c8227e3797e1b176c65b
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>