Migrate Ubuntu 16.04 docker image to common-packer
Migrate Ubuntu 16.04 docker images for netconf to the
templates available in common-packer. We are using
Ubuntu 16.04 docker images specific to GPB project
that is not maintained.

Move to the templates which are already availble in
common-packer since this will minimize the jobs and the
number of image maintained on the infra.

Tested the image with netconf CSIT job successfully:
https://logs.opendaylight.org/sandbox/vex-yul-odl-jenkins-2/netconf-csit-1node-callhome-only-fluorine/2

Change-Id: Ic3c97f2911eccbe46d1ba0d6e7ea017ae92821ad
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>