Update patch set 2
Patch Set 2:

Unfortunately, this won't work because of the priority in JJB.

https://docs.releng.linuxfoundation.org/projects/global-jjb/en/latest/best-practices.html#defaults-sections

job-groups have the highest priority when it comes to variable expansion which means we cannot define variables in this group. (I'm not sure why that {disable-job} was configured in this job-group but that variable is not effective here and instead get's parsed into a string literal `<disabled>{disable-job}</disabled>`)

The best idea I have for this situation I think would be... to create a 2nd job-group for openjdk11 (or whatever version we want to use) which hardcodes the java-version for all the jobs except the verify job (in fact probably best to exclude the verify job from the new group altogether and explicitly define it in the projects that need to test against multiple versions.

Verify really is the only job that supports multiple java-versions passed to it so we'd need to move the verify out of the new job-group so that we can define which java-versions list separately for only that job.

Patch-set: 2