Split mvn-verify jobs inte separate job-group
mvn-verify jobs expect java-version variable to be a list of strings
whereas all the other jobs just expect a single string. This becomes
problematic when used together so split them out.

This allows us to create job-groups for the odl-maven-jobs that are
JDK specific so that projects can more easily choose their release
job jdk versions.

Change-Id: I073174d6c615d3d01a771d47f281f6687d30e575
Signed-off-by: Thanh Ha <zxiiro@gmail.com>