Update patch set 1
Patch Set 1:

The interesting change is in releng-templates-java.yaml the rest of the changes are to support it. The net result of this patch should be that nothing really changes as we're just splitting out the mvn-verify jobs to a new job-group.

I'll propose a follow up patch which will create a jdk11 job group for the odl-maven-jobs group.

Patch-set: 1
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0