Update patch set 1
Patch Set 1: Code-Review-1

Why don't we fix what's broken instead of reverting the whole patch? can you link me to a job that's failing due to this?

I suspect it's caused by the "${SUITES}" and test variable in the robot run part since we had to fix that in other files.

Patch-set: 1
Label: Code-Review=-1