Update patch set 1
Patch Set 1:

Jaya, this file is auto-generated by Jenkins. Modifying it will only cause Jenkins to automatically add it back.

The only way to remove a project from autorelease is to remove the project from the pom.xml in autorelease and the git submodule by cloning the releng/autorelease project and running:

 git submodule deinit sfc
 git submodule rm sfc
 # edit pom.xml to remove sfc

Patch-set: 1