Install packages one at a time
The csit_failed_tests.txt file is not being created any
more. This fixes it.

Not sure why this breaks when installing one in a single
line, but the bs4 and requests libraries aren't found
when they are installed together with tox on the same
command line.

Also, turns out that we need bs4 now instead of
beautifulsoup4.

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>
Change-Id: Iaa45e22708cd70a358cc90b3e9a76d94becc18e3
Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>