Update patch set 1
Patch Set 1: Code-Review+1

Was troubleshooting with Allan on this and we found that the current Plastic build failures is due to JDK11 build. Switch to JDK8 until we have more time to update the code for JDK11.

Patch-set: 1
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1