Update patch set 4
Patch Set 4:

(1 comment)

> Hi Luis, I'd prefer it if you close this

You mean "close this" as in "merge this already"? :)

This change is too complicated for me to proofread reliably,
in these situations I usually switch to "just use it and see if it works" approach.
So in general, I vote for merge, just to have a proof of concept available, and fix issues later (such as double space instead of single space).
In not so general, see the comment in code; so no +1 just yet.

Patch-set: 4
Label: Verified=0