Update patch set 3
Patch Set 3:

Richard,

Given the large number of patches you currently have open and the fact that this particular patch conflicts with one of your own patches, can we please get the entire series of patches needed as a topic based set? We won't go merging in a series if any of the needed patches are failing and right now it's hard to tell which ones are or are not.

If you need help resubmitting the needed patches as a gerrit topic please ping Thanh or me on the #opendaylight-releng IRC channel.

Patch-set: 3
Label: Verified=0