Eliminate duplicate calls to delete running stack
Removed 'releng-openstack-stack-delete' macro.

This macro was left over from early development and is no longer needed.
Deleting the stack is very important. There is a cron script that will
delete orphaned agents but depending on that would result in paying for
unused resources. For more information about this change please refer to
the Issue.

Issue: RELENG-2541
Change-Id: Ia625f01e497d9688e7370eebd5eb32b0991f3b3b
Signed-off-by: Tim Johnson <tijohnson@linuxfoundation.org>