Update patch set 1
Patch Set 1: Code-Review-1

> Patch Set 1:
> 
> I think a little explanation is necessary. For the 'csit' openstack builds, the stack is getting deleted twice. Once from YAML in openstack and once from YAML in global-jjb. 
> Publishers(8): which call
>     integration-csit-publish: which calls
>          integration-csit-publish: which calls
>              releng-openstack-stack-delete: which deletes stack
>          lf-stack-delete: which deletes the stack again
> 
> releng-openstack-stack-delete: defined in Opendaylight
> lf-stack-delete: defined in global-jjb
> 
> All publishers (in jjb/integration/integration-templates.yaml) seem to make the
> same set of calls. So I deleted the call to releng-openstack-stack-delete.
> 
> I could have deleted the calls to 'lf-stack-delete' and called the 'cost' script
> from 'releng-openstack-stack-delete'. I have no idea if that is better or worse
> than what I chose.
> 
> I don't know what happens if the job completes without deleting the stack.

Can you put the explanation in the commit message? that way folks can look it up in git history.

As for why, this is legacy cruft we never got around to cleaning up. All of the stack deletes and stack related commands should come from global-jjb at this point so please do cleanup any code here. ODL was the first project to start using Stacks and most of our stack related code comes from here and most of what is in global-jjb began here in ODL.

If a job completes without deleting the stack then the stack gets orphaned in OpenStack and Jenkins doesn't know they exist and thus becomes orphaned resources wasting time and money. We MUST make sure these stacks are deleted at the end of the job so that we're not paying for resources we no longer need.

There is a cron job which runs regularly in case this mechanism fails but we should not depend on that and delete the stack in each job to make sure we minimize costs as much as possible.

Patch-set: 1
Label: Code-Review=-1