Find longterm failing jenkins jobs
Jenkins job which running on Jenkins master node
will find all jobs with consecutive failings
and put them to the file JobsWithMoreConsecutiveFailures.html
to allow examine failing jobs by clicking on hypertext links.

It is also possible to execute scripts outside of jenkins
in the order:
- bash find-failing.sh
- python find-failing.py 20
where the python script parameter is the minimal number of
consecutive failings jobs.

Signed-off-by: Stanislav Chlebec <stanislav.chlebec@pantheon.tech>
Change-Id: Idd4f8929f8a9b5bdc8127c9e1b91801ba5d4cd33