Update patch set 1
Patch Set 1: Code-Review+2

The Gerrit push macro from global-jjb handles installing of git-review:
https://github.com/lfit/releng-global-jjb/blob/eededd9e9094b6a19e2611d40c29dca30559c42b/shell/gerrit-push-patch.sh#L37-L43

But since we are not using the macro we have to handle it ourselves. It would be much more preferrable if we didn't have to relearn lessons we already learned before as writing custom code means we have to maintain custom code.

In any case this is a minor detail that we can fix in a later patch.

Patch-set: 1
Label: Code-Review=+2