Quote param values to force them to be strings
The jjb-deploy gerrit keyword that triggers the builder-jjb-deploy-job
is failing when passing a CSIT job as an argument. It was working with
another non-csit job (I tried a mvn-merge job). After some debugging,
I saw that some python lib (ElementTree.py) was throwing an exception
because a value was seen as an Int and not a string. It turned out to
be coming from our vm_x_count parameter in our csit templates. From
what I can tell, everything is now a string for those. My only guess
to why this used to work before is that some lib version has changed
and now more strict.

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>
Change-Id: If17d8e0f67e1c96f0473acd2b68ddf3aaf40d9f7