Update patch set 1
Patch Set 1:

The global-jjb update was motivated by the problems documented here:

https://jira.linuxfoundation.org/servicedesk/customer/portal/2/IT-19526

This is the error:

  19:45:06 + cp -r '.tox/*/log' '/w/workspace/lf-infra-info-master-tox-verify-any/archives/tox/*'
  19:45:06 cp: cannot stat ‘.tox/*/log’: No such file or directory 

As I understand it, in some cases tox created no log file, and the script to upload logs was not robust to a situation of no logs at all.  Full disclosure: it was my change that broke the script. 

So if you revert, then certain jobs will go back to failing.

Maybe we can figure out exactly the root cause of why logs are not uploaded, and fix that?

Patch-set: 1
Reviewer: Gerrit User 8642 <8642@7fc14799-209e-464c-9743-7a06c2c21a81>