Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > Release notes generator script is just a Bash script. I don't understand why java-version matters here or why it's being included into the job name?
> > 
> > Which part of this template uses Java and how is it used in the job?
> 
> We have two versions of this version for Sodium (jdk11 and jdk8).

We don't need 2 versions of Sodium. It is possible with JJB to setup the projects such that only 1 Sodium job is created. The release notes generator script has NOTHING to do with Java so should not be a requirement for the job.

Patch-set: 1
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>