Remove unstable bgpcep csit jobs in dist-test job
These jobs:
        bgpcep-csit-3node-bgpclustering-all
        bgpcep-csit-3node-bgpclustering-ha-only

Are always marked as ignore/not-stable when we are
vetting a release candiate. It seems the job has
been this way for multiple releases now, so likely
it's just not maintained. Removing it from the
distribution-test job so we do not have to spend
any time on it when we are working through the
release process.

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>
Change-Id: I3e55138e40c2c6f17de62e6ab5aa74e83297c1d8