Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > copy-logs: srini-netvirt-csit-3node-1cmb-0ctl-0cmp-openstack-stein-upstream-stateful-magnesium/2
> > > 
> > > Jobs failed in stein with VM error.
> > 
> > that's good data... got some bug somewhere right?
> 
> Hi Jamo,
> 
> I could not get why VMs are going to error state.
> I need a fix as stein jobs are failing and is effecting the pass count. 😞
> 
> Thanks,
> Srinivas

It's probably time to start looking for the real bug. It could be on the openstack
side. These are hard to figure out, but unless they get fixed they are really
bad for an end user to hit (imagine launching a new openstack VM for some person
but they can't connect because it did not get it's ip address).

two main places to start looking:

1) compare the flow tables on the compute and control nodes in a job that is
passing vs failing. There may be some flow missing. That was common back in the
days that I was working in netvirt.

2) search through the neutron and nova logs. Maybe there is an openstack bug
being hit and some traceback will be there?

Patch-set: 4
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>