Remove centos7-autorelease-8c-8g
A quick `find . -name "*.yaml" | xargs grep centos7-autorelease`
appears to report that this VM is no longer used by any jobs.
Let's remove it and investigate if we remove the other custom
centos7-autorelease VM as well considering autorelease is a much
smaller project today than it once was so we can likely migrate
to regular builder VMs.

Change-Id: I26f4b328497d015286c0d60ea322ccb6f76a6e41
Signed-off-by: Thanh Ha <zxiiro@gmail.com>