Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > yes but each sims in tox-verify job consumes 1Gb of RAM so it needs a least 6Gb to perform functional tests
> > thus 4c-4g has not enough RAM
> 
> That said, the jenkins minions comes without any SWAP space.
> This threshold of 6Gb might be lower if there was a swap space (but I never tried).

Ok the latest patch makes it so that only tox-verify jobs get 8c-8g and the rest of the jobs will use default 4c-4g. I confirmed with the sandbox test that this is fine for the maven job and considering tox-verify stays the same there won't be any impact there.

Patch-set: 2