Minimize CSIT dependencis in mininet-ovs-2.8 image
We are using mininet-ovs-2.8 as our tools image, which means we
are running CSIT test scripts on it. Those scripts require
the Python-2 requests library.

We used to install it before the Shell->Ansible transition via
git-review, which was part of baseline.sh. As we no longer install
it in Ansible world, neither do we install python-requests, which
is what CSIT needs.

This patch minimizes the install footprint introduced in
89cfbecea7439fae66fa10146879a9b58e71046e.

JIRA: INTTEST-100
Change-Id: I816840f83fa39592dc3361bced835e183688231e
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>