Hack to get CSIT working
something changed recently that breaks a pip install
because python2 is no longer allowed. This is doing
all the pip installs with python3 now. But, after
doing that the robot command was not in the path, so
this hard codes it to where I found it.

Clearly this is not the right/clean fix, but it
might get CSIT running again.

Also, robotframework-pycurllibrary is no longer
installed as it was failing with python3 pip. I'm
not sure it's needed in CSIT tests or not.

Signed-off-by: Jamo Luhrsen <jluhrsen@gmail.com>
Change-Id: I384bf8fda15899e40ef214c8f356f57a2ee2ad76