Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> jjb-deploy: bgpcep-csit-1node-bgp-ingest-all-magnesium

Thanh, correct me if I'm wrong, but I don't think jjb-deploy is going to use this patch to
create the job, will it? The changes here will need to be copy/pasted in to the job config.
not hard, just fyi.

I may be wrong though... let me know if I am.

Patch-set: 1