Split out release-merge job to separate project
Split all of the release-merge jobs into a separate project definition.
This resolves issue where the branch cutting script is grabbing this
definition and copying it to the next release every time we do branch
cutting causing duplicate job definitions. Considering there can be
only a single release-merge job per-project it makes sense that it
has it's own separate project definition that is separate from jobs
that can have multiple streams and branches.

Change-Id: Id701e77ed9f1ba61f6d53e4d49869d0cf118306a
Signed-off-by: Thanh Ha <zxiiro@gmail.com>