Add prettier for formatting
It appears that we inconsistently use 2 and 4 space indents for YAML
however 2 space indents appear to be the dominant one in practice so
set config to that so that the least amount of changes happen when
prettier is applied.

This patch is also a preparation patch for RELENG-152 as the automation
we want to add for managing JJB updates when branch cutting requires
a consistent formatting of the YAML files so that auto-generated output
can match appropriately with minimal formatting changes.

Issue: RELENG-152
Change-Id: Ie35521de0378b7892f06ff7f36d470847c78989f
Signed-off-by: Thanh Ha <zxiiro@gmail.com>